Dependency-Based Remote Code Repository (OOD, C++)
Under this project I essentially built a code repository which was
maintained by server and was accessible by client for code management,
it had main features of check-in and extract, and it also supported open
and closed properties of check-ins.
This repository was built to store code in package fashion and it also
stored auto generated metadata on server side which was required during
extraction by client.
HTTP style messaging communication was established between client and
server using sockets. This communication also supported sending and
receiving stream of bytes along with string messages.
GUI using WPF was also made for ease of use of repository on client
side.